backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
supabase==2.5.1
pydantic==2.12.5
pydantic-core==2.41.5
psycopg2-binary
requests>=2.31.0
gunicorn
flask-limiter==3.7.0
